refactor: move getUserConfig to node-specific util file When importing localizeUrl() into Astro components, the build failed because @proload packages were being imported in the same file, and Vite doesn't know how to transpile @proload. This moves @proload-dependent code into a separate file, so other ./utils methods can be imported in Astro code
