+++ src/__tests__/utils.test.ts
@@ -2,9 +2,7 @@ import {
   interpolate,
   createReferenceStringFromHTML,
   localizePath,
-  moveBaseLanguageToFirstIndex,
   localizeUrl,
-  deeplyStringifyObject,
   detectLocaleFromPath,
 } from "../utils";
 import i18next from "i18next";
@@ -41,22 +39,6 @@ i18next.init({
 
 afterEach(() => {
   vi.clearAllMocks();
-});
-
-describe("moveBaseLanguageToFirstIndex(...)", () => {
-  it("moves base language to first index", () => {
-    const supportedLngs = ["fr", "es", "en"];
-
-    moveBaseLanguageToFirstIndex(supportedLngs, "en");
-    expect(supportedLngs).toStrictEqual(["en", "fr", "es"]);
-  });
-
-  it("keeps base language in first index", () => {
-    const supportedLngs = ["fr", "es", "en"];
-
-    moveBaseLanguageToFirstIndex(supportedLngs, "fr");
-    expect(supportedLngs).toStrictEqual(["fr", "es", "en"]);
-  });
 });
 
 describe("interpolate(...)", () => {
@@ -364,98 +346,5 @@ describe("detectLocaleFromPath(...)", () => {
   it("with unsupported locales", () => {
     expect(detectLocaleFromPath("/de/example")).toBe("en");
     expect(detectLocaleFromPath("de/example")).toBe("en");
-  });
-});
-
-describe("deeplyStringifyObject(...)", () => {
-  it("with deep object", () => {
-    const objectToStringify = {
-      string1: "This is a string",
-      object1: {
-        string2: "This is another string",
-        string3: "This is yet another string",
-        function1: function () {
-          console.log("This is a function");
-        },
-        arrowFunction: () => {
-          console.log("This is an arrow function");
-        },
-        array1: [12346, 55, "string", { foo: "foo", bar: "bar" }],
-        object2: {
-          string4: "Again, another string!",
-        },
-        symbol1: Symbol("sym"),
-        nan1: NaN,
-        infinity1: Infinity,
-        negativeInfinity1: -Infinity,
-        undefined1: undefined,
-        null1: null,
-        boolean1: true,
-        boolean2: false,
-      },
-    };
-
-    expect(deeplyStringifyObject(objectToStringify)).toBe(
-      `{"string1": "This is a string","object1": {"string2": "This is another string","string3": "This is yet another string","function1": function() { console.log("This is a function"); },"arrowFunction": () => { console.log("This is an arrow function"); },"array1": [12346,55,"string",{"foo": "foo","bar": "bar",},],"object2": {"string4": "Again, another string!",},"symbol1": Symbol("sym"),"nan1": NaN,"infinity1": Infinity,"negativeInfinity1": -Infinity,"boolean1": true,"boolean2": false,},}`
-    );
-  });
-
-  it("with deep array", () => {
-    const arrayToStringify = [
-      {
-        string1: "This is a string",
-        object1: {
-          string2: "This is another string",
-          string3: "This is yet another string",
-          function1: function () {
-            console.log("This is a function");
-          },
-          arrowFunction: () => {
-            console.log("This is an arrow function");
-          },
-          array1: [12346, 55, "string", { foo: "foo", bar: "bar" }],
-          object2: {
-            string4: "Again, another string!",
-          },
-          symbol1: Symbol("sym"),
-          nan1: NaN,
-          infinity1: Infinity,
-          negativeInfinity1: -Infinity,
-          undefined1: undefined,
-          null1: null,
-          boolean1: true,
-          boolean2: false,
-        },
-      },
-      {
-        string1: "This is a string",
-        object1: {
-          string2: "This is another string",
-          string3: "This is yet another string",
-          function1: function () {
-            console.log("This is a function");
-          },
-          arrowFunction: () => {
-            console.log("This is an arrow function");
-          },
-          array1: [12346, 55, "string", { foo: "foo", bar: "bar" }],
-          object2: {
-            string4: "Again, another string!",
-          },
-          symbol1: Symbol("sym"),
-          nan1: NaN,
-          infinity1: Infinity,
-          negativeInfinity1: -Infinity,
-          undefined1: undefined,
-          null1: null,
-          boolean1: true,
-          boolean2: false,
-        },
-      },
-    ];
-
-    expect(deeplyStringifyObject(arrayToStringify)).toBe(
-      `[{"string1": "This is a string","object1": {"string2": "This is another string","string3": "This is yet another string","function1": function() { console.log("This is a function"); },"arrowFunction": () => { console.log("This is an arrow function"); },"array1": [12346,55,"string",{"foo": "foo","bar": "bar",},],"object2": {"string4": "Again, another string!",},"symbol1": Symbol("sym"),"nan1": NaN,"infinity1": Infinity,"negativeInfinity1": -Infinity,"boolean1": true,"boolean2": false,},},{"string1": "This is a string","object1": {"string2": "This is another string","string3": "This is yet another string","function1": function() { console.log("This is a function"); },"arrowFunction": () => { console.log("This is an arrow function"); },"array1": [12346,55,"string",{"foo": "foo","bar": "bar",},],"object2": {"string4": "Again, another string!",},"symbol1": Symbol("sym"),"nan1": NaN,"infinity1": Infinity,"negativeInfinity1": -Infinity,"boolean1": true,"boolean2": false,},},]`
-    );
   });
 });
